% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNactivation.R
\name{activation}
\alias{activation}
\title{The three functions used in neural networks (sigmoidal, linear or tanh)}
\usage{
activation(x, arg = "tanh")
}
\arguments{
\item{x}{The inputs of the hidden neuron.}

\item{arg}{The option of one of the two training functions (tanh by default)}
}
\value{
The	hidden neuron output z.
}
\description{
The three functions used in neural networks (sigmoidal, linear or tanh)
}

