% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelChoice.R
\name{choose_kernel}
\alias{choose_kernel}
\title{Defines the desired kernel for our problem.}
\usage{
choose_kernel(u, arg = NULL)
}
\arguments{
\item{u}{A    number/vector.}

\item{arg}{The optional kernel (Alphabetical)}
}
\value{
The	kernelized values.
}
\description{
Defines the desired kernel for our problem.
}
\note{
If argument is not alphabetical, it returns the Gaussian Kernel
}

