% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetric.R
\name{distance_metric}
\alias{distance_metric}
\title{Defines the desired metric for our problem.}
\usage{
distance_metric(x, y, arg = "Eucledean", p = NULL)
}
\arguments{
\item{x}{A    number/vector.}

\item{y}{A    number/vector.}

\item{arg}{The optional distance metric (Alphabetical/Eucledean by default)}

\item{p}{For the Lp-norm (requires arg='L')}
}
\value{
The	distance	of	x	and	y.
}
\description{
Defines the desired metric for our problem.
}
\note{
If argument is not alphabetical, it returns the Eucledean Metric
}

