% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkernel_matrix.R
\name{gkernel_matrix}
\alias{gkernel_matrix}
\title{The general Kernel Matrix}
\usage{
gkernel_matrix(x, h, distance = "Eucledean", kernel = "Gaussian")
}
\arguments{
\item{x}{The observations}

\item{h}{The bandwidth}

\item{distance}{The distance metric}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The k((x[i]-x[j])/h) matrix with i rows and j columns
}
\description{
The general Kernel Matrix
}
\note{
By default we use the Gaussian Kernel
}

