% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelMatrix.R
\name{kernel_matrix}
\alias{kernel_matrix}
\title{The Kernel Matrix}
\usage{
kernel_matrix(x, h, kernel = "Gaussian")
}
\arguments{
\item{x}{The observations}

\item{h}{The bandwidth}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The k((x[i]-x[j])/h) matrix with i rows and j columns
}
\description{
The Kernel Matrix
}
\note{
By default we use the Gaussian Kernel
}

