% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNNreg.R
\name{kNNreg}
\alias{kNNreg}
\title{The k-nearest neighborhood regression}
\usage{
kNNreg(y, x, k = 1, weight = FALSE, method = "Gaussian")
}
\arguments{
\item{y}{The regressant}

\item{x}{The regressor}

\item{k}{The number of neighbors (1 neighbor by default)}

\item{weight}{Assigns different weights to neighbors according to distance}

\item{method}{If weight=TRUE then this is the kernel function to evaluate the weights}
}
\description{
The k-nearest neighborhood regression
}

