% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npquant.R
\name{npquant}
\alias{npquant}
\title{Nonparametric Quantile Regression}
\usage{
npquant(y, x, a = 0.5, hx = 0.05, hy = 0.1, method = "Gaussian")
}
\arguments{
\item{y}{The regressant}

\item{x}{The regressor}

\item{a}{Which quantile}

\item{hx}{The bandwidth for xs}

\item{hy}{The bandwidth for ys}

\item{method}{For the kernel function}
}
\description{
Nonparametric Quantile Regression
}

