% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParzenRosenblatt.R
\name{parzen_rosenblatt}
\alias{parzen_rosenblatt}
\title{The Parzen-Rosenblatt density estimator}
\usage{
parzen_rosenblatt(x, h, kernel = "Gaussian")
}
\arguments{
\item{x}{The observations}

\item{h}{The bandwidth}

\item{kernel}{The alphabetical argument for the kernel function of choice (from choose_kernel)}
}
\value{
The estimated f for every observation
}
\description{
The Parzen-Rosenblatt density estimator
}
\note{
By default we use the Gaussian Kernel
}

